Update 50_TelegramBot.pm
I added readings for the client's message time. This is useful for identifying old or delayed messages due to connection problems or other errors.
For example:
FHEM commands that execute critical routines like disarming the alarm system or unlocking a door can be ignored if the message was sent too long ago.